Adding comments, general small improvements/better implementations

+++ CoreJavaProject/src/Q15Interface.java
@@ -1,8 +1,8 @@
 
 public interface Q15Interface {
 
-	int addition(int a, int b);
-	int subtraction(int a, int b);
-	int multiplication(int a, int b);
-	int division(int a, int b);
+	double addition(double a, double b);
+	double subtraction(double a, double b);
+	double multiplication(double a, double b);
+	double division(double a, double b);
 }
